# Asset-Organizer/lib/api-zod/src/generated/types/cityPopulationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Populle World Population API
 * OpenAPI spec version: 0.1.0
 */
import type { CityData } from "./cityData";

export interface CityPopulationResponse {
  data: CityData[];
  year: number;
  totalCities: number;
}