# Asset-Organizer/lib/api-zod/src/generated/types/countryData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Populle World Population API
 * OpenAPI spec version: 0.1.0
 */

export interface CountryData {
  iso3: string;
  name: string;
  continent: string;
  /** Population in thousands */
  population: number;
  populationMillions: number;
  /** People per square km */
  density?: number;
  lat?: number;
  lon?: number;
  year: number;
  growthRate?: number;
}