# Asset-Organizer/lib/api-zod/src/generated/types/timeseriesPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Populle World Population API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeseriesPoint {
  year: number;
  population: number;
  populationMillions: number;
  isProjection: boolean;
}