# Asset-Organizer/lib/api-zod/src/generated/types/timeseriesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Populle World Population API
 * OpenAPI spec version: 0.1.0
 */
import type { LocationTimeseries } from "./locationTimeseries";

export interface TimeseriesResponse {
  locations: LocationTimeseries[];
  variant: string;
}